backend/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
frozendict==2.3.8
html5lib==1.1
idna==3.4
lxml==4.9.3
multitasking==0.0.11
numpy==1.26.0
pandas==2.1.1
peewee==3.16.3
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.5
webencodings==0.5.1
yfinance==0.2.30
